# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/unit-variants/{id}": {
      "get": {
        "tags": [
          "unit-variant-controller"
        ],
        "operationId": "getUnitVariantById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnitVariants"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "unit-variant-controller"
        ],
        "operationId": "updateUnitVariant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitVariants"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnitVariants"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "unit-variant-controller"
        ],
        "operationId": "deleteUnitVariant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/properties/{id}": {
      "get": {
        "tags": [
          "property-details-controller"
        ],
        "operationId": "getPropertyById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "property-details-controller"
        ],
        "operationId": "updateProperty",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "property-details-controller"
        ],
        "operationId": "deleteProperty",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pricing-details/{id}": {
      "get": {
        "tags": [
          "pricing-details-controller"
        ],
        "operationId": "getPricingDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PricingDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "pricing-details-controller"
        ],
        "operationId": "updatePricingDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PricingDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PricingDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "pricing-details-controller"
        ],
        "operationId": "deletePricingDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/payment-plans/{id}": {
      "get": {
        "tags": [
          "payment-plans-controller"
        ],
        "operationId": "getPaymentPlan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentPlans"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "payment-plans-controller"
        ],
        "operationId": "updatePaymentPlan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentPlans"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentPlans"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "payment-plans-controller"
        ],
        "operationId": "deletePaymentPlan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/unit-variants": {
      "get": {
        "tags": [
          "unit-variant-controller"
        ],
        "operationId": "getAllUnitVariants",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnitVariants"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "unit-variant-controller"
        ],
        "operationId": "createUnitVariant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitVariants"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnitVariants"
                }
              }
            }
          }
        }
      }
    },
    "/api/properties": {
      "get": {
        "tags": [
          "property-details-controller"
        ],
        "operationId": "getAllProperties",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyDetails"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "property-details-controller"
        ],
        "operationId": "createProperty",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/pricing-details": {
      "get": {
        "tags": [
          "pricing-details-controller"
        ],
        "operationId": "getAllPricingDetails",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PricingDetails"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "pricing-details-controller"
        ],
        "operationId": "createPricingDetail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PricingDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PricingDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/payment-plans": {
      "get": {
        "tags": [
          "payment-plans-controller"
        ],
        "operationId": "getAllPaymentPlans",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentPlans"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "payment-plans-controller"
        ],
        "operationId": "createPaymentPlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentPlans"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentPlans"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PropertyDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "propertyType": {
            "type": "string"
          },
          "developer": {
            "type": "string"
          },
          "totalTowers": {
            "type": "integer",
            "format": "int32"
          },
          "totalUnits": {
            "type": "integer",
            "format": "int32"
          },
          "totalFloors": {
            "type": "integer",
            "format": "int32"
          },
          "possessionDate": {
            "type": "string",
            "format": "date"
          },
          "unitVariants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnitVariants"
            }
          }
        }
      },
      "UnitVariants": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "propertyDetails": {
            "$ref": "#/components/schemas/PropertyDetails"
          },
          "unitType": {
            "type": "string"
          },
          "sizeSqFt": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number"
          }
        }
      },
      "PaymentPlans": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "pricingDetails": {
            "$ref": "#/components/schemas/PricingDetails"
          },
          "planType": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "PricingDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "basePricePerSqFt": {
            "type": "number"
          },
          "maintenanceChargesPerMonth": {
            "type": "number"
          },
          "bookingAmount": {
            "type": "number"
          },
          "paymentPlans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentPlans"
            }
          }
        }
      },
...
    }
  }
}